Moved eval_answer from tools to core
